backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client and Web Scraping
httpx>=0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0

# NLP and LLM
spacy==3.7.2
ollama==0.1.6
anthropic>=0.40.0

# Excel Export
openpyxl==3.1.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# CORS
python-multipart==0.0.6
